Chore: refactor build
* Breaking change
* Removes CJS and UMD builds in favor of ES2015
