+++ .storybook/main.cjs
@@ -1,10 +1,7 @@
 module.exports = {
-  stories: [
-    '../stories.js',
-  ],
   addons: [
     '@storybook/addon-a11y',
     '@storybook/addon-controls',
   ],
-  framework: '@storybook/react',
+  stories: ['../stories/index.tsx'],
 }
