+++ package.json
@@ -2,8 +2,11 @@
   "name": "react-medium-image-zoom",
   "version": "4.4.0",
   "description": "Accessible medium.com-style image zoom for React",
-  "main": "dist/cjs/index.js",
-  "module": "dist/esm/index.js",
+  "main": "dist/index.js",
+  "exports": {
+    ".": "./dist/index.js"
+  },
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "git@github.com:rpearce/image-zoom.git"
@@ -59,23 +62,17 @@
     "build": "run-s clean build:js build:css build:docs",
     "build:css": "cp ./source/styles.css ./dist/styles.css",
     "build:docs": "mkdir -p docs && build-storybook -o docs",
-    "build:js": "rollup -c ./rollup.config.js",
-    "build:js:watch": "npm run build:js -- -w",
-    "ci": "run-p lint build test",
-    "clean": "run-p clean:dist clean:docs",
-    "clean:dist": "rm -rf ./dist",
-    "clean:docs": "rm -rf ./docs",
+    "build:js": "tsc",
+    "ci": "run-p lint test build",
+    "clean": "rm -rf ./dist && rm -rf ./docs",
     "contributors:add": "all-contributors add",
     "contributors:generate": "all-contributors generate",
-    "coverage": "jest --coverage --coverageReporters=text-lcov | coveralls",
-    "lint": "eslint . --ext .js,.ts,.tsx",
-    "prepublishOnly": "run-p lint test && npm run build:js",
+    "lint": "eslint .",
+    "prepublishOnly": "run-p lint test build",
     "start": "start-storybook -p 6006",
     "test": "jest"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^22.0.0",
-    "@rollup/plugin-node-resolve": "^13.3.0",
     "@storybook/addon-a11y": "^6.4.22",
     "@storybook/addon-controls": "^6.4.22",
     "@storybook/react": "^6.4.22",
@@ -90,8 +87,6 @@
     "@typescript-eslint/parser": "^5.23.0",
     "all-contributors-cli": "^6.20.0",
     "axe-core": "^4.4.2",
-    "babel-loader": "^8.2.5",
-    "coveralls": "^3.1.1",
     "eslint": "^8.15.0",
     "eslint-plugin-jest": "^26.1.5",
     "eslint-plugin-jsx-a11y": "^6.5.1",
@@ -103,9 +98,6 @@
     "npm-run-all": "^4.1.5",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
-    "rollup": "^2.72.1",
-    "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-typescript2": "^0.31.2",
     "ts-jest": "^28.0.2",
     "typescript": "^4.6.4"
   },
