+++ tsconfig.json
@@ -6,15 +6,14 @@
     "esModuleInterop": true,
     "importHelpers": true,
     "jsx": "react",
-    "lib": ["es2020", "dom"],
-    "module": "es6",
+    "lib": ["esnext", "dom"],
+    "module": "esnext",
     "moduleResolution": "node",
     "noImplicitAny": true,
     "outDir": "./dist/",
-    "sourceMap": true,
+    "sourceMap": false,
     "strict": true,
-    "target": "es5"
+    "target": "es2015"
   },
-  "exclude": ["node_modules/"],
-  "include": ["./source/**/*", "./__tests__/**/*"]
+  "include": ["./source/index.tsx"]
 }
